Still working through pony tutorial
My current thought is that I can have two Entities:

class NameList(Entity):
  title = Required(str)
  items = Set("Items")

class Items(Entity):
  value = Required(str)
  name_list = Set(NameList)

So that one item can be in multiple lists.

The flow is:

A list is added, and respective entries are created.
Then, if another list is added, if it has duplicates, the list they are
now also a part of can be added to its list, instead of duplicating
items.

It also means creating combined lists is very easy.

I do want date added, but not sure if that'd be useful for anything, and
I don't know how to have multiple date added's for when an item is added
to multiple lists.

There will be a few internal lists, so that things like "Which names
have I generated?" is easy to keep track of.

I want an undo history, so that modificatinos to items can be staged,
and then either committed, or not. Should also make recovering from
crashes cleaner, if every modification is added to a queue in the
database, in synchrony with UI updates, and so if the app crashes, it
can resume cleanly.