Trying to add initial file import functionality
Really had no idea where to start with this application, so I started
with functionality I know I want it to have.

I know I want a GUI, so I could try designing from that first, and stub
out various functions for filling in later, which would probably help
with figuring out which functionality needs to be implemented.

However, I knew for a fact it needs to be able to do the following:

- Import a text file, taking each item as a name
- Exporting a list as a text file

Those actions will need to be implemented, regardless of the way the
interface is laid out, and in order to make them "intuitive", they need
to support things like:

- Adding the same file multiple times, with different contents each time
- Adding a file with the same contents as another file, but renamed
  - This should not create an entirely new, duplicate list
- Exporting a list to an existing file by appending new entries

I think that by meeting those requirements, I'll end up adding a
significant portion of the functionality I want the application to have.

Currently, I don't even know what's really achievable easily, and which
operations and idioms will be fast, and this I think will greatly
influence the design of the GUI part of the application.

For example: if the idiom of "Show me the order in which names were
added to this list" is slow, then that would greatly affect how sorting
is presented in the appllication.

I'm also extremely unfamiliar with databases in general, and PonyORM as
the middleman, so implementing this functionality will definitely help
build familiarity with the concepts necessary for this application.