Added preliminary Pony models
Have to finish reading the pony docs to learn how to use it properly.
I have a feeling that the Name entity may be able to be replaced with an
index or something.

Also, still not sure how I want to represent these models in the
database interface, or how I am going to get the types sorted so mypy
believes me when I say I can use a function that takes Any and returns a
definite type at runtime.
Probably lots of isinstance() checking.

Also, mypy.ini was checking against Python 3.7 rules, while the Pipfile
was using 3.8. That's fixed.