Following tkinter tutorials
https://tkdocs.com/tutorial/firstexample.html
https://youtu.be/YXPyB4XeYLA

I think it's behoove me to implement the graphical interfaces, and  as I
go along, add necessary functionality to database_interface, like custom
classes with metaclasses.

I think I could probably implement everything except the saving, and
stub that out, to make sure the interfaces work, before solidifying
everything in the database.

I also want to make sure I'm not avoiding building a particular part of
the app, just because I think making changes to it after it's built will
be tough, since that usually means the structure I have in mind could be
modified to make refactoring easier.

I think this is the route to go, though: design the gui, then the
database interface, then the database.