Scrollable frame expands; need class
So, as a class, it needs an obvious way to support adding objects to,
and configuring the grid properties of the inner, scrollable frame.

I also _really_ want it to handle automatically showing and hiding the
scrollbars as needed, both for vertical and horizontal scrolling.

Not sure how well I'll be able to do that for text, but we'll see.