Started TDD
Still need tests for whether database does or doesn't create a file if
it is/isn't told to.

Thakfully, PonyORM Database objects store the entities attached to them,
so I can wrap the entire database initialization into a function and use
the single returned database object to find the associated Entity
classes.