Finished adding tests for current database
This by no means means that it's complete.

It literally only initializes a database.

It's still missing a "Deleted" ItemList to hold deleted items

It still needs to implement all of the CRUD methods:

- Create
- Read
- Update
- Delete

Once those are implemented, a corresponding set of classes can be
created in database_interface that use those methods to implement
actually saving the data.