Added pydantic as dependency
pydantic looks like it'll help me greatly accelerate the creation of
classes that mirror pony's Entity classes:

It has an ORM more:
https://pydantic-docs.helpmanual.io/usage/models/#orm-mode-aka-arbitrary-class-instances

It also can generate a JSON Object for each "model" object, and can
parse an incoming one, meaning as long as there's a Lua
parser/generator, it'll be way easy to make transfer data.

fastapi will be used to implement the REST API for getting the
LOVE-based GUI up and running.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch dev
# Your branch is up to date with 'origin/dev'.
#
# Changes to be committed:
#	modified:   Pipfile.lock
#	modified:   mypy.ini
#	new file:   namepicker/pydantic_learning.py
#	modified:   poetry.lock
#	modified:   pyproject.toml
#